+++ app/components/shared/back_confirm_dialog.js
