+++ app/screens/CareerPlanning/contact_screen.ios.js
@@ -17,7 +17,7 @@ import FooterBar from '../../components/FooterBar';
 import { FontSetting } from "../../assets/style_sheets/font_setting";
 
 import mainStyles from '../../assets/style_sheets/main/main';
-import BackConfirmDialog from '../../components/back_confirm_dialog';
+import BackConfirmDialog from '../../components/shared/back_confirm_dialog';
 import SchoolListView from '../../components/schools/school_list';
 
 import realm from '../../schema';
