+++ app/screens/CareerPlanning/personality_jobs_screen.android.js
@@ -10,7 +10,7 @@ import {
 } from 'react-native';
 import Toast, { DURATION } from 'react-native-easy-toast'
 
-import BackConfirmDialog from '../../components/back_confirm_dialog';
+import BackConfirmDialog from '../../components/shared/back_confirm_dialog';
 import CheckboxGroup from '../../components/checkbox_group';
 import FooterBar from '../../components/FooterBar';
 import MaterialIcon from 'react-native-vector-icons/MaterialIcons';
