+++ app/screens/CareerPlanning/personality_jobs_screen.ios.js
@@ -11,7 +11,7 @@ import {
 import Toast, { DURATION } from 'react-native-easy-toast'
 import { Divider } from 'react-native-elements';
 
-import BackConfirmDialog from '../../components/back_confirm_dialog';
+import BackConfirmDialog from '../../components/shared/back_confirm_dialog';
 import CheckboxGroup from '../../components/checkbox_group';
 import FooterBar from '../../components/FooterBar';
 import MaterialIcon from 'react-native-vector-icons/MaterialIcons';
