+++ app/screens/CareerPlanning/personality_screen.ios.js
@@ -12,7 +12,7 @@ import {
 
 import Toast, { DURATION } from 'react-native-easy-toast';
 
-import BackConfirmDialog from '../../components/back_confirm_dialog';
+import BackConfirmDialog from '../../components/shared/back_confirm_dialog';
 import CloseButton from '../../components/shared/close_button';
 
 import mainStyles from '../../assets/style_sheets/main/main';
