+++ app/screens/CareerPlanning/recommendation_screen.ios.js
@@ -8,7 +8,7 @@ import {
   BackHandler,
 } from 'react-native';
 
-import BackConfirmDialog from '../../components/back_confirm_dialog';
+import BackConfirmDialog from '../../components/shared/back_confirm_dialog';
 import FooterBar from '../../components/FooterBar';
 
 import mainStyles from '../../assets/style_sheets/main/main';
