+++ app/screens/CareerPlanning/summary_screen.android.js
@@ -13,7 +13,7 @@ import Toast, { DURATION } from 'react-native-easy-toast';
 
 import CheckboxGroup from '../../components/checkbox_group';
 import RadioButtonGroup from '../../components/radio_button_group';
-import BackConfirmDialog from '../../components/back_confirm_dialog';
+import BackConfirmDialog from '../../components/shared/back_confirm_dialog';
 import FooterBar from '../../components/FooterBar';
 import MaterialIcon from 'react-native-vector-icons/MaterialIcons';
