+++ app/screens/admin/dashboard_screen.js
@@ -12,7 +12,7 @@ import {
 import * as Progress from 'react-native-progress';
 import AwesomeIcon from 'react-native-vector-icons/FontAwesome';
 import MaterialIcon from 'react-native-vector-icons/FontAwesome';
-import Button from '../../components/button';
+import Button from '../../components/shared/button';
 
 // Utils
 import realm from '../../schema';
