+++ app/screens/profile/profile_form.js
@@ -21,7 +21,7 @@ import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view
 
 // Components
 import InputTextContainer from '../../components/input_text_container';
-import SaveButton from '../../components/save_button';
+import SaveButton from '../../components/shared/save_button';
 import PickerSpecific from '../../components/picker/PickerSpecific';
 
 import FamilySituation from '../../data/json/family_situation.json';
