+++ app/screens/school/institution_stack.ios.js
@@ -13,7 +13,7 @@ import { FontSetting } from '../../assets/style_sheets/font_setting';
 
 import Filter from '../../components/schools/filter';
 import BackButton from '../../components/shared/back_button';
-import SaveButton from '../../components/save_button';
+import SaveButton from '../../components/shared/save_button';
 
 import SchoolScreen from './school_screen';
 import InstitutionDetail from './institution_detail';
