+++ app/screens/stackNav/career_counsellor_stack.ios.js
@@ -7,7 +7,7 @@ import store from '../../redux/store';
 import headerStyles from '../../assets/style_sheets/header';
 import CloseButton from '../../components/shared/close_button';
 import NextButton from '../../components/NextButton';
-import SaveButton from '../../components/save_button';
+import SaveButton from '../../components/shared/save_button';
 import BackButton from '../../components/shared/back_button';
 
 import PersonalUnderstandingForm from '../PersonalUnderstanding/PersonalUnderstandingForm';
